requirements.txt
# Core Deep Learning Frameworks
tensorflow==2.10.0
keras==2.10.0

# Web Framework
Django==3.2.20

# Image Processing
opencv-python==4.8.0.74
Pillow==9.5.0

# Data Processing
numpy==1.23.5
pandas==1.5.3

# Scientific Computing
scikit-learn==1.2.2
scipy==1.10.1

# Visualization
matplotlib==3.7.1

# Jupyter (for notebook)
jupyter==1.0.0
notebook==6.5.4
ipykernel==6.23.1

# Additional utilities
h5py==3.8.0
